Lecture 124 - building & using a custom context provider component.
Refactored the logging in logic into auth-context store so that
components that need the isLoggedIn state can use it through the
Context API instead of through props.

We can place logic that is used is many different components in one
centralized location by creating a component that encapsulates a context
provider such as:
  const MyContext = React.createContext({
    someState: initialValue,
    someFunction: (args) => {}
  });

  export const MyContextProvider = (props) => {
    const [myState, setMyState] = React.useState({initial state});
    const someFunction = (newState) => {
      setMyState(newState);
    };

    return (
      <MyContext.Provider value={{
        myState: myState,
        someFunction: someFunction
      }}>
        {props.children}
      </MyContextProvider>
    );
  };

  export default MyContext;

Components anywhere in the component tree that need data stored in this
context can import this file and call the useContext hook to access
the data:
  import MyContext from "fileWithContext.js";

  const SomeComponent = (props) => {
    const myContext = React.useContext(MyContext);
    const someState = myContext.someState;
    const someFunction = myContext.SomeFunction;

    return (<div></div>);
  };

The lowest common ancestor of all components that require some context
data should be wrapped with the custom provider component.
  // in index.js (if all components need some context)
  ...
  root.render(
    <MyContextProvider>
      <App />
    </MycontextProvider>
  );